Dump agent stdout and stderr directly to file (don't pipe to console)
