+++ aries-backchannels/acapy/acapy-main.env
@@ -1,3 +1,6 @@
 
 LOG_LEVEL=DEBUG
 RUST_LOG=warn
+# set PIPE_AGENT_OUTPUT=True to pipe output from agents "live"
+# (default "False" will dump agent stdout and stderr to separate files)
+# PIPE_AGENT_OUTPUT=True
