+++ aries-backchannels/python/utils.py
@@ -242,3 +242,14 @@ def pad_base64(value: str) -> str:
         n_missing = 4 - (len(value) % 4)
         value = value + (n_missing * "=")
     return value
+
+
+def str2bool(v):
+    if isinstance(v, bool):
+        return v
+    if v.lower() in ("yes", "true", "t", "y", "1"):
+        return True
+    elif v.lower() in ("no", "false", "f", "n", "0"):
+        return False
+    else:
+        raise argparse.ArgumentTypeError("Boolean value expected.")
